Pre-tiled and gdal-metadata handling is not WOfS-specific.
